revised my_alarm struct to contain an array for the command to make reading/writing the structure eaiser, also worked on the client a bunch, specifically the add function.
I (plan) to revies the command array to contain just a filename with
either a tmp file with a shell script to run or a path to an audio file
to run.

I also started to add syslog calls into the daemon itself, only errors
for now but hopefully I'll put in more.