Cleaned up various issues to get everything to compile, started adding code to make alarmd systemd compatible
I probably would've written things differently if I thought to use
systemd from the start, but as it is now I'm just going to make sure
that things work with systemd, but also will work without it.

I still need to grind out the rest of the code for the frontend, most of
it should be easy, just argument parsing and then communicating with the daemon.