added a bunch of assembly files
Some are just files copied from as, as referances
but test.S and start_stub.s are assembly files that in theory I'll be
able to assemble and link without any other files and run as an
executable. In theory it should reverse each command line argument and
print the reversed arguments to stdout. It doesn't work yet, but I'll
need to use gdb to fix it and I don't feel like that right now