Updated life_sdl to use multiple threads, and cleaned up the code
Currently there are only 2 threads, one which updates the grid and
pixels and one which runs the event loop and updates the screen.

I broke the code for life_sdl into 3 seperate files.

I made quite a few other minor chages in the process of doing this, but
nothing too major.

I currently use pthreads since I throught I was going to use signals and
sdl threads can't do that, but I don't so I'm going to convent to sdl
threads.