Rearranged vndb.scm and added some threading
Specifically I load the cache/tags/traits files in a seperate thread,
since it takes a few seconds, which is noticable when loading the
file. I don't do any synchronization so it's not safe, but it shouldn't
be a problem as long as I don't access the variables within the first
couple seconds.

Also I added a simple front end which just runs a repl when the file is
run as a script