Fixed some compiler errors and added a simple logger class
I got things to run but ran into an error dealing whith handling throttling
error, so I decided to add a logger to make it eaiser to see what was going
wrong (I was using the wrong field to check the error type). I haven't tested
the program yet, but it compiles fine.