Fixed some issues with jpeg rendering
Specifically I was using SDL_RGB888 as my texture format when I needed
to use SDL_RGB24, This was a pain in the ass to figure out.

I wrote a test program to read jpegs from the database and display them and it
works, but viewing them interactively doesn't. I think it may have somethnig to
do with threads so I need to rework my test program to run SDL in a seperate
thread and see if that breaks it.