Fixed a bug in expand_variable which broke variables containing '_'
I was finding the end of a variable by searching for a non-alphanumeric
character. Now I find the end by searching for a '[' or the end of the
string. This assumes the string that gets passed to expand_variable doesn't
have any weird characters in it, and I'm not sure I do enough validatition to
ensure that this is true so that's something I need to fix.